Obtaining Match info from yarascanner collect
I think that would be useful to add Match info in yarascanner collect method:
https://groups.google.com/forum/#!topic/rekall-discuss/eNldxgjfFbM

Before this addition, when you use collect() method in order to scan a dump, you lose the match info associated to the rule.

But now, it returns the complete match info. I think that is a small change that do not affect the actual behaviour of rekall plugins. 

Moreover will be useful to the developers that there are using rekall as API.